JwtFilter to SecurityConfiguration
